Move controls into tool level
